+++ managers/webpubManifest.py
@@ -13,15 +13,21 @@ class WebpubManifest:
         self.openSection = None
 
     # TODO validate kwargs against schema.org/Book
-    def addMetadata(self, dcdwRecord):
+    def addMetadata(self, dcdwRecord, conformsTo=None):
         self.metadata['title'] = dcdwRecord.title
 
         if len(dcdwRecord.authors) > 0:
             self.metadata['author'] = list(dcdwRecord.authors[0].split('|'))[0]
 
-        isbns = list(filter(lambda x: x[-4:] == 'isbn', dcdwRecord.identifiers))
+        isbns = list(filter(
+            lambda x: x[-4:] == 'isbn', dcdwRecord.identifiers
+        ))
+
         if len(isbns) > 0:
             self.metadata['identifier'] = 'urn:isbn:{}'.format(isbns[0][:-5])
+
+        if conformsTo:
+            self.metadata['conformsTo'] = conformsTo
 
     def addSection(self, sectionTitle, sectionURL):
         if self.openSection:
@@ -36,7 +42,7 @@ class WebpubManifest:
     def closeSection(self):
         if self.openSection:
             self.tableOfContents.append(self.openSection)
-        
+
         self.openSection = None
 
     # TODO Make it possible to add subsections iteratively
@@ -71,7 +77,10 @@ class WebpubManifest:
             'metadata': self.metadata,
             'links': self.links,
             'readingOrder': self.readingOrder,
-            'resources': [{'href': res, 'type': 'application/pdf'} for res in self.resources],
+            'resources': [
+                {'href': res, 'type': 'application/pdf'}
+                for res in self.resources
+            ],
             'toc': self.tableOfContents
         }