+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 2021-08-03 -- v0.8.0
 ### Added
 - New endpoint `utils/proxy` to allow for proxying of resources to webreader
+- `conformsTo` parameter for Webpub Manifests for identifying PDF resources
 
 ## 2021-07-21 -- v0.7.1
 ### Fixed